modification des roles

+++ src/main/java/com/example/fmw/repository/IRoleRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.fmw.entity.Role;
 
-public interface IRoleRepository extends JpaRepository<Role, Long> {
+public interface IRoleRepository extends JpaRepository<Role, Integer> {
     Role findByName(String name);
 }
